MasterImproved seems to be up and runnnnnniinnnnnggggg!

+++ com/drin/java/analysis/clustering/Clusterer.java
@@ -9,9 +9,9 @@ import java.util.List;
 
 public interface Clusterer {
 
-   public Map<Double, List<Cluster>> getClusters();
+   public Map<Float, List<Cluster>> getClusters();
    public String getName();
-   public double getInterClusterSimilarity();
+   public float getInterClusterSimilarity();
 
    public void clusterData(List<Cluster> clusters);
    public void setProgressCanvas(JTextArea canvas);
