+++ com/drin/java/metrics/IsolateSimpleMetric.java
@@ -9,9 +9,9 @@ import java.util.Map;
 import java.util.HashMap;
 
 public class IsolateSimpleMetric extends DataMetric<Isolate> {
-   protected Map<String, Map<String, Map<String, Double>>> mRegionMap;
+   protected Map<String, Map<String, Map<String, Float>>> mRegionMap;
 
-   public IsolateSimpleMetric(Map<String, Map<String, Map<String, Double>>> corrMap) {
+   public IsolateSimpleMetric(Map<String, Map<String, Map<String, Float>>> corrMap) {
       super();
 
       mRegionMap = corrMap;
@@ -20,7 +20,7 @@ public class IsolateSimpleMetric extends DataMetric<Isolate> {
    @Override
    public void apply(Isolate elem_A, Isolate elem_B) {
       for (String region : mRegionMap.keySet()) {
-         Map<String, Map<String, Double>> corrMap = mRegionMap.get(region);
+         Map<String, Map<String, Float>> corrMap = mRegionMap.get(region);
 
          /*
          System.out.printf("elem_A: '%s' elem_B: '%s'\n", elem_A.getName(), elem_B.getName());
@@ -31,13 +31,13 @@ public class IsolateSimpleMetric extends DataMetric<Isolate> {
 
          if (corrMap.containsKey(elem_A.getName()) &&
              corrMap.get(elem_A.getName()).containsKey(elem_B.getName())) {
-            Map<String, Double> tmp_map = corrMap.get(elem_A.getName());
+            Map<String, Float> tmp_map = corrMap.get(elem_A.getName());
 
             //System.out.printf("%s\n", elem_A.getName());
 
             if (tmp_map.containsKey(elem_B.getName())) {
                //System.out.printf("%s\n", elem_B.getName());
-               mResult += tmp_map.get(elem_B.getName()).doubleValue();
+               mResult += tmp_map.get(elem_B.getName()).floatValue();
             }
             else {
                setError(-1);
@@ -46,14 +46,14 @@ public class IsolateSimpleMetric extends DataMetric<Isolate> {
          }
 
          else if (corrMap.containsKey(elem_B.getName()) &&
-             corrMap.get(elem_B.getName()).containsKey(elem_A.getName())) {
-            Map<String, Double> tmp_map = corrMap.get(elem_B.getName());
+                  corrMap.get(elem_B.getName()).containsKey(elem_A.getName())) {
+            Map<String, Float> tmp_map = corrMap.get(elem_B.getName());
 
             //System.out.printf("else %s\n", elem_B.getName());
 
             if (tmp_map.containsKey(elem_A.getName())) {
                //System.out.printf("%s\n", elem_A.getName());
-               mResult += tmp_map.get(elem_A.getName()).doubleValue();
+               mResult += tmp_map.get(elem_A.getName()).floatValue();
             }
             else {
                setError(-1);
@@ -68,8 +68,8 @@ public class IsolateSimpleMetric extends DataMetric<Isolate> {
    }
 
    @Override
-   public double result() {
-      double result = mResult / mRegionMap.size();
+   public float result() {
+      float result = mResult / mRegionMap.size();
 
       //Logger.debug(String.format("comparison: [%.05f]", result));
       this.reset();
